requirements.txt
moviepy==1.0.3
requests==2.32.4
python-dotenv==1.1.1
pydantic-settings
replicate